tests: add read length check tests
The length of the data produced by the server and sent to the caller is
recorded and compared to ensure that the data length is identical along
the path.

This test identified a path where too much data was returned from the
server (it is no security issue, but rather a performance issue). This
is now fixed.

Signed-off-by: Stephan Mueller <smueller@chronox.de>