initialize DRNG only with 256 bits of entropy
Remove the initialization of 32 / 128 bits of entropy stages. The code
for handling them is left in case we want to go back on it.

The reason is a suggestion from Joshua E. Hill mentioning that an
SP800-90A DRBG must be fully seeded during instantiation to be compliant
with the stipulation of 90A.

Signed-off-by: Stephan Mueller <smueller@chronox.de>