esdm: add support for interrupt handling
The read() client operation is interrupted every 512ms which ensures
that if the server does not answer somehow (e.g. it blocks on the
request), a consuming application can terminate the read request
without sending a signal, but by registering a function callback that
indicates whether to terminate the read operation.

This implies that the client RPC library API is updated as follows:

- esdm_rpcc_init_unpriv_service / esdm_rpcc_init_priv_service receive a
  new parameter that is a function that is invoked to return whether the
  caller wants the operation to terminate. This may be NULL in case the
  caller does not want to register such callback.

- all API calls received an API call with the void *int_data allowing
  the caller to provide a pointer to some arbitrary memory buffer that
  is used to invoke the interrupt callback. The old APIs are still
  present and work as they used to.

- add an interrupt handler in the CUSE code that uses the aforementioned
  framework. This now allows terminating a client on /dev/random or
  /dev/urandom with a signal. Without this code, terminating a blocking
  client is impossible.

Update the test code to compile.

Signed-off-by: Stephan Mueller <smueller@chronox.de>