esdm: CUSE devices - serialize privileged operations
Considering that CUSE allows parallel requests, the CUSE privileged
operation handling must ensure that before the privilege escalation is
performed, all other calls are completed.

This is achieved by using a reader/writer lock. The reader lock is used
for unprivileged calls, the writer lock is used for privileged calls.

Signed-off-by: Stephan Mueller <smueller@chronox.de>