esdm: simplify PR handling code
In case SP800-90C compliance is required, the random number gathering
loop now can fully satisfy this request. This prevents an overuse of the
entropy sources in case of generating more data with PR enabled than 32
bytes. The overusing is due to forcing a reseed and then terminating
this loop. By re-entering the forced reseed is honored followed by the
PR reseed. Now, the forced reseed is not applied when in PR mode

Signed-off-by: Stephan Mueller <smueller@chronox.de>