esdm: fix deadlocks during shutdown
Various deadlocks are caused by the following situation: a thread that
just took a lock receives a signal and thus terminates without releasing
the lock. The cleanup thread takes the lock to ensure exclusive access
to the resource causing a deadlock. This is fixed by using a robust lock
which returns if the thread that locked the lock terminates.

Signed-off-by: Stephan Mueller <smueller@chronox.de>