esdm: move blocking behavior into client
When invoking a blocking interface, the ESDM server (or rather the
library) returns -EAGAIN instead of blocking. This error code is used to
implement a poll-block on the client side. This ensures that connections
are not left open and the ESDM server can be shut down gracefully in
this case.

Signed-off-by: Stephan Mueller <smueller@chronox.de>