esdm: add forced seeding to reach fully seeded level
If the ESDM cannot obtain sufficient entropy to reach the fully seeded
level, the entropy sources are pulled for data repeadetly. As this is an
atomic operation, the entropy can be summed up to reach the fully seeded
level.

This new function also implies that the interfaces contain callbacks to
indicate whether new "entropy" is available for triggering the
mechanism.

Signed-off-by: Stephan Mueller <smueller@chronox.de>