esdm: simplify forced seeding implementation
Revert the changes in esdm_es_mgr.* regarding the forced seeding
implementation. The forced seeding is only relating to the DRNG
management and is unrelated to the entropy sources. Thus, it must not be
present in esdm_es_mgr.*.

The reimplementation uses a trigger called by the interface functions
which triggers the forced seeding process.

Signed-off-by: Stephan Mueller <smueller@chronox.de>