esdm: apply forced seeding to atomic DRNG
The forced seeding support was made available to the DRNGs feeding user
space. This patch makes this support available to the atomic DRNG which
feeds kernel space consumers. This implies that as soon as a fast
entropy source becomes available even with less than perfect entropy,
the atomic DRNG is attempted to be fully seeded using the forced seeding
operation.

To achieve that behavior, the following changes are applied:

- adding lrng_get_atomic which makes the atomic DRNG handle available to
  the DRNG manager

- remove lrng_drng_atomic_seed_es considering that the atomic DRNG seeding
  from entropy sources uses the forced reseeding logic implemented in the
  DRNG manager.

- change lrng_drng_seed_es_nolock to only invoke lrng_init_ops when the
  user space DRNGs are seeded as the result of the init_ops call only
  applies to the user space interfaces

- change lrng_drng_seed_es to not call out to the atomic DRNG seeding
  any more

- change __lrng_drng_seed_work to add the atomic DRNG seeding from
  entropy sources in case the LRNG is not yet initialized (and thus
  the user space DRNGs are available to seed the atomic DRNG).

- Fast entropy sources of CPU and Jitter RNG are changed to invoke the
  lrng_force_fully_seeded function to trigger the forced seeding for
  whatever DRNG is available at the time these ES initialize.

In addtion, considering the forced seeding works as designed, it is applied
as soon as the first request to random numbers is received and the respective
DRNG is not yet fully seeded. Thus, the "waiting" for 5 requests for random
numbers is removed. This is considered appropriate because the entropy
sources uses in the forced seeding are defined to provide the entropy
credited to them. Therefore it makes no real sense to wait until slow
entropy sources may have collected entropy in the mean time. Besides, if a
user is not satisfied with the state, it can force a reseeding of the
DRNGs using a write to /dev/random. This change is visible in
lrng_force_fully_seeded which invokes __lrng_drng_seed_work right away.

Signed-off-by: Stephan Mueller <smueller@chronox.de>