esdm: safeguard for min_t and max_t - fix compilation
The macros are no problem when a and b are variables or constants,
but if they are functions, the functions are potentially executed twice.
If those functions will yield different results at first and second
execution, the macro wil not work as intended.

This patch uses inlines for the differnt types which is seemingly the
only way for ISO-C as groups are not allowed.

Signed-off-by: Stephan Mueller <smueller@chronox.de>