fix: allow dissimilar ES buffer sizes between kernel and ESDM
The entropy buffer struct entropy_es contains the entropy buffer e which
has a size that depends on the compile time configuration. This very same
buffer exists in the kernel code for the IRQ/Sched ES as well. However,
it is possible that both the kernel and user space have now different
buffer sizes which need to be handled in a safe manner. The buffer
sizes are allowed to be as follows:

- 32 bytes - this applies when there is no oversampling configured (e.g.
  the FIPS 140-related oversampling is not in effect)

- 32 + 16 bytes - the oversampling is in effect.

This leads to the following scenarios that are now handled by the ESDM
ES handlers for both the Sched and IRQ ES - those scenarios are detected
during initialization of the user space ES handlers:

- kernel and ESDM have the same buffer -> the ESDM can now write directly
  into the entropy buffer provided by the ES manager

- kernel has larger buffer (i.e. 48 bytes) than ESDM (has 32 bytes) ->
  considering that the kernel fills that buffer with a hash function,
  entropy is spread evenly in that buffer -> fill only 32 bytes of the
  ES manager-provided buffer with the kernel data and reduce the kernel-
  reported entropy rate proportionally

- kernel has smaller buffer (i.e. 32 bytes) than ESDM (has 48 bytes) ->
  the ES manager-provided buffer is filled with all kernel data and the
  entropy rate is set to the kernel-reported value. Yet, the kernel
  can ever at most deliver only 32 bytes of entropy which may imply that
  this ES alone can never cause the ESDM to reach full entropy.

This logic is implemented in a common service function used by the
Sched and IRQ ES handlers.

In addition, this patch removes the code duplication for setting the
requested entropy size. This change includes a fix that prevents setting
this value over and over again even though the kernel has this value
already.

Reported-by: Markus Theil
Signed-off-by: Stephan Mueller <smueller@chronox.de>