Sched ES: when active, IRQ ES is claimed to have zero bits of entropy
If the scheduler-based entropy source is enabled, the IRQ ES is claimed
to not return any entropy. This is due to the fact that interrupts may
trigger scheduling events. This implies that interrupts are not
independent of interrupts.

This handling is identical as implemented for the Kernel RNG ES in
esdm_es_krng.

Signed-off-by: Stephan Mueller <smueller@chronox.de>