Terminate accept() syscall with signal
During shutdown, the accept() syscall must be unblocked. This is ensured
by sending an uncaught signal which terminates the respective thread.

This terminates the unprivileged handler thread. The privileged handler
thread does not need to be terminated considering that it executes in the
main thread of the process. This main thread handles the signal of the
shutdown request and thus already does not wait for the accept any more.

Signed-off-by: Stephan Mueller <smueller@chronox.de>