fix test
Due to the patch that added the forced (re)seeding, the test implementation
failed as an entropy source with non-zero entropy was present. This
source is used to fully seed the DRNG again which implied that the DRNG
never reached the unseeded state.

By setting all entropy sources to have zero bits of entropy, the test now
can again validate that the DRNG goes back to unseeded mode.

Reported-by: Markus Theil
Signed-off-by: Stephan Mueller <smueller@chronox.de>