IRQ/Sched ES: retry to access kernel
When the ESDM server command line options -s or -i are used, the
ES retry to access the kernel in case initial access cannot be made.
The goal is that in case the boot sequence races between the ESDM server
and the insmod of the esdm_es.ko, the boot sequence still allows the
ESDM server to access the kernel in the end.

Yet, the retry is only done with the given command line flags to avoid
an indefinite retry when it is known that the kernel support will
never become online.

Signed-off-by: Stephan Mueller <smueller@chronox.de>