Jitter RNG: enable buffer monitor
The Jitter RNG buffer monitor fills a buffer with output from the Jitter
RNG on a separate thread, the monitor thread. When trying to get random
data, the buffer is simply read. At intervals, the buffer is refilled
again by the thread. The goal is to ensure that callers do not need to
wait for the generation of the Jitter RNG processing.

The buffer support can be disabled at compile time as well as disabled
at runtime - the first case, the extra memory required for the operation
is not compiled either.

Signed-off-by: Stephan Mueller <smueller@chronox.de>