Jent ES: block management - remove lock, pull 256 bits only
The lock was a left-over from issues faced during development.

The Jitter RNG is now pulled for only 256 bits of entropy which is
the common value at runtime. As the buffer support shall speed up the
runtime, the potential oversampling during initialization time requires
a synchronous call to the Jitter RNG.

Finally, wake up the monitor thread during shutdown to ensure it receives
the request to terminate the Jitter RNG handler.

Signed-off-by: Stephan Mueller <smueller@chronox.de>