add kernel-based jitter source
In our measurements with AARCH64-based NICs (Nvidia Bluefield 2),
the userspace jitter RNG provided much less entropy than the
kernel-based jitter RNG. The reason for this seems to be the
high-resolution timestamp cntvct_el0, which only has a resolution
of 100 MHz, while the CPU of the NIC clocks at 2.5 GHz. The
non-virtualized timestamp is not accesible from user-space.
The high resolution PMU timestamp is also not directly accesible from
user-space.

As we did not want to spare CPU cycles for the internal
timer-thread in jitterentropy, our current proposal is to add this
optional kernel-based jitterentropy source.
This source has access to a high-resolution timestamp and provides
a higher entropy count in our test samples. (Most likely it uses
cntpct_el0).

Signed-off-by: Markus Theil <theil.markus@gmail.com>