RPC server: wait for monitor thread
The monitor thread is started and then the start sequence of the server
waits for the initialization immediately after the thread launch.
There is a possibility to enter a deadlock otherwise.

Also, only kill the non-special threads at the end as the special
threads are killed at the server shutdown.

Signed-off-by: Stephan Mueller <smueller@chronox.de>