Jitter RNG: Instantiate async Jitter RNG
Due to the removed locking in the async instance, it is possible that the
synchronous call in esdm_jent_async_get is operating at the same time as
the Jitter RNG ES monitor of esdm_jent_async_monitor. This is not allowed
because the Jitter RNG is required to operate signle-threaded. Thus, to
still not require a lock, use a 2nd Jitter RNG instance.

Signed-off-by: Stephan Mueller <smueller@chronox.de>