catch theoretical out of lock situation
According to the documentation, it is possible that too many locks are
taken at the same time. In this case, the pthread_rwlock_rdlock will
return EAGAIN. Catch it to avoid indicating to the caller that the
lock is erroneously taken.

This change covers an issue not observed in the wild yet.

Signed-off-by: Stephan Mueller <smueller@chronox.de>