CUSE: wakeup select/poll during suspend/sleep
When the system goes into suspend, the processes doing a select/poll must
be woken up as otherwise the suspend will not complete.

This is achieved by allowing the ESDM server to receive a SIGUSR1 which
in turn notifies the CUSE daemons about that signal. The SIGUSR1 is triggered
by a systemd unit which fires when systemd enters a suspend/sleep cycle.

Signed-off-by: Stephan Mueller <smueller@chronox.de>