CUSE: use atomic variable for shutdown
Only the use of an atomic variable guarantees the propagation of a
variable change to all threads immediately when the variable is set.

This fixes a potential deadlock where the CUSE daemon is supposed to
terminate under high load: Due to the high load, the poll loop may
started again before the shutdown is propagating which implies that
the CUSE daemon was hanging in a sem_wait call.

Signed-off-by: Markus Theil <theil.markus@gmail.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>