increase performance of RPC
The RPC used to be implemented by sending one RPC data object at a time.
However, all data objects can also be copied into a linear buffer and
then send with one write system call.

However, this implies that a new buffer of the size of the data to be
sent must be allocated which effectively duplicates the amount of
memory required for an RPC call. Commonly this duplication should be
harmless because one call can at most be 64k Bytes.

This patch implements the linearization of the data when the compile-time
option of small_memory is disabled (the default). If the ESDM server
is used on systems with memory pressure, small_memory can be enabled and
the old behavior is compiled.

With the linearization of the buffer, the performance of the following
command is increased from about 70MB/s to 170MB/s on my test system.

dd if=/dev/random of=/dev/null bs=65520 count=10000

Signed-off-by: Stephan Mueller <smueller@chronox.de>