Add RPC calls for entropy status
For NTG.1 generator usage of ESDM monitoring the
overall entropy count in ESDM is also of interest besides
the entropy count in the auxiliary pool.

When waiting for ESDM to become fully seeded for the
first time in the boot sequence, the other two RPC calls
allow writing a small daemon with minimal effort for this
use case.

Signed-off-by: Markus Theil <theil.markus@gmail.com>