Kernel Jent ES: add proper locking
As the ES is used by a multi-threaded server, add a lock around the
use of the kernel AF_ALG handle. The lock only needs to ensure that
the handle is not allocated/deallocated while another process is
accessing it. The runtime use of the handle does not need any
serialization as the kernel performs this. Therefore an R/W lock
is used to take the reader lock when using the handle during runtime
and the write lock when modifying the handle.

Signed-off-by: Stephan Mueller <smueller@chronox.de>