fix: prevent RPC DoS when using slow operations
The DRBG with PR and access to the seed data are a slow operation. If
multiple parallel requests are received, they can fill up the avaiable
RPC links. Thus, limit access to the slow operations to one request at
any given time. Additional requests are returned -EAGAIN which is
received by the client lib to restart the request again.

Signed-off-by: Stephan Mueller <smueller@chronox.de>