handle rogue clients more gracefully.
If a client received a notification to supply entropy, but it fails to
send anything, the ESDM will not send a notification again. This issue
is now alleviated by checking the need_entropy common variable before
sleeping on the semaphore allowing another caller to identify whether
entropy is needed.

The implementaton is racy in the sense that between the check for
need_entropy and sleeping on the semaphore the semaphore could fire and
the rogue client consumes the semaphore trigger. For now, we accept this
issue because the client is intended to react on the notification. It
may be fixed by a sem_timedwait or similar to regularly check
for the need_entropy.

Signed-off-by: Stephan Mueller <smueller@chronox.de>