rpc_client: better track and close stale connections
After experiencing client timeouts on connections already closed by the
server, better tracking of this is added.

Furthermore, preventively close RPC connections in the client, which
are probably already closed by the server.

Should the new timeout of 2s be too low or high in some use-cases,
it can easily be configured now.

While there, zeroize the write buffer and improve the comment
on additional measures for this to be effective.

Signed-off-by: Markus Theil <theil.markus@gmail.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>