esdm: select -> ppoll
With the recent increase of default worker threads
to 1024 select can no longer be used, as it can
only handle 1024 fds. With at least one additional
server fd for accept we are getting over this limit
now. As we mostly only wait with a small amount of
sockets, ppoll is easier to use than epoll and fast
enough.

Signed-off-by: Markus Theil <theil.markus@gmail.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>