es_aux: rework backtracking operation
Before this patch, the aux source re-inserted its output
in the hash state for backtracking resistance.

When evaluating this source in isolation, it would be possible
to predict future outputs if a single output is known and
further extractions used no or only a small amount of fresh input
entropy to the aux pool(s).

Fix this, by introducing an internal state, which is never output
and use it for backtracking resistance updates.

Signed-off-by: Markus Theil <markus.theil@secunet.com>